pyproject.toml
[project]
name = "google-map-api-multi-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.8.0",
    "google-maps-places>=0.2.2",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "pyright>=1.1.403",
]
